Removed warnings and fixed issue regarding app creating more no of tickets than user entered(Removed action from textfield)

+++ Housie/PdfCreator.m
@@ -107,7 +107,7 @@
                     no = [no stringByAppendingString:@" "];
                 }
             }
-            length = no.length;
+            length = (int)no.length;
             const char* no1 = [no UTF8String];
             CGContextShowTextAtPoint(pdfContext, nox, noy, no1, length);
             if (index == 8 || index == 17 ) {
